requirements.txt
Fabric==1.10.0
PyYAML==3.11
ecdsa==0.11
paramiko==1.15.1
pycrypto==2.6.1
zope.interface==4.1.1
